pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hal-9000"
version = "0.1.0"
description = "AI-powered research assistant for PDF processing, knowledge organization, and experiment design"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Bryan Wisk" }
]
keywords = [
    "research",
    "pdf",
    "obsidian",
    "knowledge-management",
    "ai",
    "claude",
    "materials-science"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "pdfplumber>=0.10.0",
    "pypdf>=4.0.0",
    "sqlalchemy>=2.0.0",
    "anthropic>=0.40.0",
    "rich>=13.0.0",
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    # Acquisition module dependencies
    "httpx>=0.27.0",
    "aiofiles>=23.0.0",
    "tenacity>=8.2.0",
    # Gateway dependencies
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.3.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
]
gdrive = [
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
]
vector = [
    "chromadb>=0.5.0",
]
docs = [
    "mkdocs-material>=9.5.0",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
hal = "hal9000.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/hal9000"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/config",
    "/templates",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
timeout = 30
timeout_method = "thread"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
